feat: 3단계 요구사항 - 예약 추가 API 구현

+++ src/main/java/roomescape/Reservation.java
@@ -1,19 +1,19 @@
 package roomescape;
 
 public class Reservation {
-    private int id;
+    private long id;
     private String name;
     private String date;
     private String time;
 
-    public Reservation(int id, String name, String date, String time) {
+    public Reservation(long id, String name, String date, String time) {
         this.id = id;
         this.name = name;
         this.date = date;
         this.time = time;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
